add ability to dynamically update the reset voltage
The shuttering async calls should be "tracking", i.e. they shouldn't remember an old outdated target voltage since that might change in the meantime. So if the global led state changes while a tracking async call is waiting, it'll update the target voltage to whatever the new global state is.